ABle to get groups displayed on map and display group info when info window is clicked
- modified group class for case when groupMembers list is empty
- added empty groupMembers list condition check in populateListView

+++ app/src/main/java/ca/cmpt276/walkinggroup/dataobjects/MyItem.java
@@ -7,7 +7,7 @@ public class MyItem implements ClusterItem {
     private final LatLng mPosition;
     private final String mTitle;
     private final String mSnippet;
-    private int grpId;
+    private Long grpId;
 
     public MyItem(double lat, double lng) {
         mPosition = new LatLng(lat, lng);
@@ -17,7 +17,7 @@ public class MyItem implements ClusterItem {
 
     }
 
-    public MyItem(double lat, double lng, String title, String snippet, int grpId) {
+    public MyItem(double lat, double lng, String title, String snippet, Long grpId) {
         mPosition = new LatLng(lat, lng);
         mTitle = title;
         mSnippet = snippet;
@@ -39,5 +39,5 @@ public class MyItem implements ClusterItem {
         return mSnippet;
     }
 
-    public int getGrpId(){ return grpId;}
+    public Long getGrpId(){ return grpId;}
 }
